refactor : MVC 패턴을 고려한 검증 책임
