+++ src/main/java/racingcar/controller/TurnsGenerator.java
@@ -1,33 +1,8 @@
 package racingcar.controller;
 
-import racingcar.view.RacingCarError;
-
 public class TurnsGenerator {
 
-    private static final int ZERO_VALUE = 0;
-
     public static int generateTurn(String candidateNumber) {
-        checkTurns(candidateNumber);
         return Integer.parseInt(candidateNumber);
-    }
-
-    public static void checkTurns(String turnCandidate) {
-        int integerCandidate = checkInteger(turnCandidate);
-        checkNegative(integerCandidate);
-    }
-
-    private static int checkInteger(String turnCandidate) {
-        try {
-            return Integer.parseInt(turnCandidate);
-        } catch (Exception error) {
-            RacingCarError.notInteger();
-            return ZERO_VALUE;
-        }
-    }
-
-    private static void checkNegative(int integerCandidate) {
-        if (integerCandidate < ZERO_VALUE) {
-            RacingCarError.negativeInteger();
-        }
     }
 }
